+++ CHANGELOG.md
@@ -2,3 +2,5 @@
 
 # v0.0.8 (24/01/2023)
 - Added the `sample_rate` parameter.
+- Updated the `energy_threshold` parameter to be a `float` or a `(float,
+  float)`. This allows to parameter a fixed length recording.